Module powerfmt::buf

A buffer for constructing a string while avoiding heap allocation.

Structs

WriteBuffer

A buffer for construct a string while avoiding heap allocation.

Functions

maybe_uninit_slice_assume_init_mut 🔒 ⚠

Equivalent of MaybeUninit::slice_assume_init_mut that compiles on stable.

maybe_uninit_slice_assume_init_ref 🔒 ⚠

Equivalent of MaybeUninit::slice_assume_init_ref that compiles on stable.

maybe_uninit_uninit_array 🔒

Equivalent of MaybeUninit::uninit_array that compiles on stable.

maybe_uninit_write_slice 🔒

Equivalent of MaybeUninit::write_slice that compiles on stable.