Crate powerfmt

powerfmt is a library that provides utilities for formatting values. Specifically, it makes it significantly easier to support filling to a minimum width with alignment, avoid heap allocation, and avoid repetitive calculations.

Modules

buf

A buffer for constructing a string while avoiding heap allocation.

ext

Extension traits.

smart_display

Definition of SmartDisplay and its related items.