powerfmt::ext

Trait FormatterExt

pub trait FormatterExt: Sealed {
    // Required method
    fn pad_with_width(&mut self, width: usize, args: Arguments<'_>) -> Result;
}

An extension trait for core::fmt::Formatter.

Required Methods

fn pad_with_width(&mut self, width: usize, args: Arguments<'_>) -> Result

Writes the given arguments to the formatter, padding them with the given width. If width is incorrect, the resulting output will not be the requested width.