Struct powerfmt::smart_display::FormatterOptions

pub struct FormatterOptions { /* private fields */ }

Configuration for formatting.

This struct is obtained from a Formatter. It provides the same functionality as that of a reference to a Formatter. However, it is not possible to construct a Formatter, which is necessary for some use cases of SmartDisplay. FormatterOptions implements Default and has builder methods to alleviate this.

Implementations

impl FormatterOptions

pub fn with_fill(&mut self, c: char) -> &mut Self

Sets the fill character to use whenever there is alignment.

pub fn with_sign_plus(&mut self, b: bool) -> &mut Self

Set whether the + flag is specified.

pub fn with_sign_minus(&mut self, b: bool) -> &mut Self

Set whether the - flag is specified.

pub fn with_align(&mut self, align: Option<Alignment>) -> &mut Self

Set the flag indicating what form of alignment is requested, if any.

pub fn with_width(&mut self, width: Option<usize>) -> &mut Self

Set the optional integer width that the output should be.

pub fn with_precision(&mut self, precision: Option<usize>) -> &mut Self

Set the optional precision for numeric types. Alternatively, the maximum width for string types.

pub fn with_alternate(&mut self, b: bool) -> &mut Self

Set whether the # flag is specified.

pub fn with_sign_aware_zero_pad(&mut self, b: bool) -> &mut Self

Set whether the 0 flag is specified.

impl FormatterOptions

pub const fn fill(&self) -> char

Character used as โfillโ whenever there is alignment.

pub const fn align(&self) -> Option<Alignment>

Flag indicating what form of alignment was requested.

pub const fn width(&self) -> Option<usize>

Optionally specified integer width that the output should be.

pub const fn precision(&self) -> Option<usize>

Optionally specified precision for numeric types. Alternatively, the maximum width for string types.

pub const fn sign_plus(&self) -> bool

Determines if the + flag was specified.

pub const fn sign_minus(&self) -> bool

Determines if the - flag was specified.

pub const fn alternate(&self) -> bool

Determines if the # flag was specified.

pub const fn sign_aware_zero_pad(&self) -> bool

Determines if the 0 flag was specified.