powerfmt::smart_display

Struct Metadata

pub struct Metadata<'a, T>
where
    T: SmartDisplay + ?Sized,
{ /* private fields */ }

Information used to format a value. This is returned by SmartDisplay::metadata.

This type is generic over any user-provided type. This allows the author to store any information that is needed. For example, a typeโs implementation of SmartDisplay may need to calculate something before knowing its width. This calculation can be performed, with the result being stored in the custom metadata type.

Note that Metadata always contains the width of the type. Authors do not need to store this information in their custom metadata type.

Generally speaking, a type should be able to be formatted using only its metadata, fields, and the formatter. Any other information should be stored in the metadata type.

Implementations

impl<'a, T> Metadata<'a, T>
where T: SmartDisplay + ?Sized,

pub const fn new( unpadded_width: usize, _value: &T, metadata: T::Metadata, ) -> Self

Creates a new Metadata with the given width and metadata. While the width should be exact, this is not a requirement for soundness.

pub fn reuse<'b, U>(self) -> Metadata<'b, U>
where U: SmartDisplay<Metadata = T::Metadata> + ?Sized, 'a: 'b,

Reuse the metadata for another type. This is useful when implementing SmartDisplay for a type that wraps another type. Both typeโs metadata type must be the same.

pub const fn unpadded_width(&self) -> usize

Obtain the width of the value before padding.

pub fn padded_width(&self, f: FormatterOptions) -> usize

Obtain the width of the value after padding.

impl Metadata<'_, Infallible>

pub fn unpadded_width_of<T>(value: T, f: FormatterOptions) -> usize
where T: SmartDisplay,

Obtain the width of the value before padding, given the formatter options.

pub fn padded_width_of<T>(value: T, f: FormatterOptions) -> usize
where T: SmartDisplay,

Obtain the width of the value after padding, given the formatter options.